requirements.txt
# Python 3.11
kafka-python
tkintermapview
matplotlib
numpy
psycopg
